[New] Added test cases for transitions referrals details
